client code to test line server performance
client code to test line server performance.  This compiles and runs separately.   The way to start it is:

go run client.go  <ip-address-of-lineserver>:<port-number-of-line-server>

eg.:

if line server is running in a server with an address 10.20.30.40 reachable by client, and the port the server is listening is 12345, then the command is:

go run client.go 10.20.30.40:12345